+++ Broker.cpp
@@ -1,9 +1,9 @@
 #include <string>
 
-class StockBrockerDriver {
-public:
-	virtual bool login(std::string id, std::string passwd) = 0;
-	virtual bool buy(std::string code, int price, int quantity) = 0;
-	virtual bool sell(std::string code, int price, int quantity) = 0;
-	virtual int getPrice(std::string code) = 0;
-};
+//class StockBrockerDriver {
+//public:
+//	virtual bool login(std::string id, std::string passwd) = 0;
+//	virtual bool buy(std::string code, int price, int quantity) = 0;
+//	virtual bool sell(std::string code, int price, int quantity) = 0;
+//	virtual int getPrice(std::string code) = 0;
+//};
