+++ autotradingsystem.cpp
@@ -1,5 +1,8 @@
 #include <string>
 
+#include "Broker.h"
+#include "MockDriver.h"
+
 class AutoTradingSystem {
 public:
 	void selectStockBroker(std::string blockername) {
