+++ autotradingsystem.h
@@ -1,0 +1,11 @@
+#include <string>
+
+class AutoTradingSystem {
+public:
+    std::string broker_name;
+    // std::unique_ptr<StockBrockerDriver> broker;
+
+    bool selectStockBroker(std::string broker_name);
+
+    std::string getCurrentBrokerName();
+};
