+++ kiwer_adapter.cpp
@@ -1,5 +1,5 @@
 #include "kiwer_api.cpp"
-#include "Broker.cpp"
+#include "Broker.h"
 #include <string>
